POSHI-609 Move PoshiProperties to poshi core
