+++ modules/test/poshi/poshi-core/src/main/java/com/liferay/poshi/core/PoshiGetterUtil.java
@@ -16,7 +16,6 @@ import com.liferay.poshi.core.util.ExternalMethod;
 import com.liferay.poshi.core.util.FileUtil;
 import com.liferay.poshi.core.util.GetterUtil;
 import com.liferay.poshi.core.util.OSDetector;
-import com.liferay.poshi.core.util.PoshiProperties;
 import com.liferay.poshi.core.util.PropsUtil;
 import com.liferay.poshi.core.util.StringUtil;
 import com.liferay.poshi.core.util.Validator;
