+++ modules/test/poshi/poshi-core/src/main/java/com/liferay/poshi/core/elements/PoshiNode.java
@@ -5,8 +5,8 @@
 
 package com.liferay.poshi.core.elements;
 
+import com.liferay.poshi.core.PoshiProperties;
 import com.liferay.poshi.core.script.PoshiScriptParserException;
-import com.liferay.poshi.core.util.PoshiProperties;
 import com.liferay.poshi.core.util.StringUtil;
 
 import java.net.URL;
