+++ modules/test/poshi/poshi-core/src/main/java/com/liferay/poshi/core/util/Dom4JUtil.java
@@ -5,6 +5,8 @@
 
 package com.liferay.poshi.core.util;
 
+import com.liferay.poshi.core.PoshiProperties;
+
 import java.io.CharArrayWriter;
 import java.io.IOException;
 import java.io.StringReader;
